添加批量删除，批量开启，批量关闭，状态广播，全局异常处理

+++ src/main/java/cn/mrcsh/zfmcserverpanelapi/entity/structure/Container.java
@@ -61,6 +61,10 @@ public class Container {
 
     public void shutdown() {
         log.info("正在关闭实例:{}",containerId);
+        if(this.getStopCmd() != null){
+            sendCommand(this.getStopCmd());
+            return;
+        }
         this.process.destroyForcibly();
         try {
             if (errorStream != null) errorStream.close();
